+++ AudioKit/Common/Internals/Soundpipe/modules/padsynth.c
@@ -28,8 +28,8 @@ int sp_gen_padsynth(sp_data *sp, sp_ftbl *ps, sp_ftbl *amps,
 {
 
     int i, nh;
-    int N = ps->size;
-    int number_harmonics = amps->size;
+    int N = (int) ps->size;
+    int number_harmonics = (int) amps->size;
     SPFLOAT *A = amps->tbl;
     SPFLOAT *smp = ps->tbl;
